Add: Main class. Fix: syntax
